pyproject.toml
[project]
name = "duo-live"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
]
